+++ src/main/java/net/partyaddon/access/GroupManagerAccess.java
@@ -5,4 +5,6 @@ import net.partyaddon.group.GroupManager;
 public interface GroupManagerAccess {
 
     public GroupManager getGroupManager();
+
+    public void setGroupManager(GroupManager groupManager);
 }