+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - 
 ### Fixed:
 - Fixed xaeroplus compat
+- Fixed party issue on player death
 ### Changed:
 -